// src/app/protected/settings/settings.interface.ts
export interface Pixel {
    index: any;
    channelId: any;
    pixelId: number;
    pixelName: string;
    pixelView: string;
    pixelCustom: string;
    pixelCustomId: number;
    pixelCustomEvent: string;
    pixelFBId: string;
    pixelComment: string;
    pixelStatus: boolean;
    pixelCreated: any;
    pixelLastupdate: any;
}